# backend/internal/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Run:ai RCA Backend API",
    "version": "0.1.56",
    "description": "Versioned HTTP contract for the Run:ai RCA Backend. Unless an operation states otherwise, successful responses use the envelope { \"status\": \"ok\", \"data\": ... } and list endpoints add a \"pagination\" block."
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
...
    "/api/v1/knowledge-candidates": {
      "get": {
        "tags": ["Knowledge"],
        "summary": "List knowledge candidates",
        "parameters": [{ "name": "status", "in": "query", "description": "Filter by lifecycle status (e.g. ready_for_review).", "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "Incident-derived knowledge candidates",
            "content": { "application/json": { "examples": { "candidates": { "$ref": "#/components/examples/KnowledgeCandidateListEnvelope" } } } }
          }
        }
      }
    },
...
    "/api/v1/knowledge-packages": {
      "get": {
        "tags": ["Knowledge"],
        "summary": "List knowledge packages",
        "parameters": [{ "name": "include_retired", "in": "query", "schema": { "type": "boolean", "default": false } }],
        "responses": {
          "200": {
            "description": "Active (and optionally retired) packages",
            "content": { "application/json": { "examples": { "packages": { "$ref": "#/components/examples/KnowledgePackageListEnvelope" } } } }
          }
        }
      }
    },
...
    "/api/v1/knowledge/families": {
      "get": {
        "tags": ["Knowledge"],
        "summary": "List configured root-cause families",
        "responses": {
          "200": {
            "description": "Root-cause family catalog",
            "content": { "application/json": { "example": { "status": "ok", "data": { "families": ["gpu_hardware_error", "scheduler_preemption", "storage_csi_attach", "network_cni", "oom_kill"] } } } }
          },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "Error": { "type": "object", "properties": { "error": { "type": "string" } } },
...
    },
    "examples": {
...
      "KnowledgeCandidateListEnvelope": {
        "value": {
          "status": "ok",
          "data": [{ "candidate_id": "kc_44ab19", "status": "ready_for_review", "title": "GPU node NotReady after driver upgrade", "root_cause_family": "gpu_hardware_error", "kind": "signature", "confidence": 0.86, "incident_id": "inc_5f3a9c2b", "supporting_case_count": 3, "created_at": "2026-07-16T09:19:00Z" }],
          "pagination": { "total": 9, "limit": 20, "offset": 0, "has_more": false }
        }
      },
...
      "KnowledgePackageListEnvelope": {
        "value": {
          "status": "ok",
          "data": [{ "package_id": "kp_12ab34", "status": "active", "title": "GPU node NotReady after driver upgrade", "root_cause_family": "gpu_hardware_error", "candidate_id": "kc_44ab19", "confidence": 0.86, "published_at": "2026-07-16T10:05:00Z", "mirror_status": "synced", "runtime_status": "active" }],
          "pagination": { "total": 5, "limit": 20, "offset": 0, "has_more": false }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Error response",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "incident not found" } } }
      }
    }
  }
}